requirements.txt
PyYAML==6.0.1
GitPython==3.1.40
sentence-transformers==2.2.2
openai==1.3.8
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1
openclaw>=0.1.0
clawstr>=0.1.0
